# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Swagger recipes
  description: Our recipes's API
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/roikatz16/recepies_API/1.0.0
  description: SwaggerHub API Auto Mocking
- url: https://localhost:4000
...
paths:
  /auth/Register:
    post:
      tags:
      - auth
      description: A JSON Object contains username, password, and another details.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body'
        required: true
      responses:
        "201":
          description: |
            user created succssesfully
        "409":
          description: |
            Username taken
  /auth/Login:
    post:
      tags:
      - auth
      description: A JSON Object contains username, password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_1'
        required: true
      responses:
        "200":
          description: |
            Login succeeded The session ID is returned in a cookie named 'session' You need to include this cookie in subsequent request
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: session=e8sd828hnskzx12w
        "400":
          description: Username or Password incorrect
        "401":
          description: No Username in the system
...
components:
  schemas:
...
    body:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: Contains only letters
        password:
          maxLength: 10
          minLength: 5
          type: string
          description: Contains letters with at least 1 numerical character and 1
            scecial character
        first_name:
          type: string
        last_name:
          type: string
        country:
          type: string
        email:
          type: string
      example:
        username: roi
        password: 1234%p
        first_name: roi
        last_name: levi
        country: Japan
        email: roi@gmail.com
    body_1:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: Contains only letters
        password:
          maxLength: 10
          minLength: 5
          type: string
          description: Contains letters with at least 1 numerical character and 1
            scecial character
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: session
      in: cookie